// frontend/src/api/remoteControlApi/model.ts
export interface DeviceData {
	[key: string]: number | string | undefined;
	reg_33?: number;
	reg_35?: number;
	reg_36?: number;
	reg_38?: number;
	reg_64?: number;
	reg_65?: number;
	reg_66?: number;
	reg_68?: number;
	reg_71?: number;
	reg_75?: number;
	reg_76?: number;
	reg_77?: number;
	reg_78?: number;
	reg_96?: number;
	reg_97?: number;
	reg_99?: number;
	reg_108?: number;
	reg_109?: number;
	reg_110?: number;
	reg_111?: number;
	reg_128?: number;
	reg_133?: number;
	reg_192?: number;
	reg_193?: number;
	reg_195?: number;
	reg_257?: number;
	reg_258?: number;
	reg_260?: number;
	reg_512?: number;
	reg_608?: number;
	reg_610?: number;
	reg_640?: number;
	reg_646?: number;
	reg_673?: number;
	reg_674?: number;
	reg_675?: number;
	reg_676?: number;
	reg_677?: number;
	reg_678?: number;
	reg_679?: number;
	reg_680?: number;
	reg_681?: number;
	reg_685?: number;
	reg_704?: number;
	reg_705?: number;
	reg_707?: number;
	reg_708?: number;
	reg_736?: number;
	reg_737?: number;
	reg_739?: number;
	reg_740?: number;
	reg_741?: number;
	reg_745?: number;
	reg_746?: number;
	reg_834?: number;
	she_hum?: number;
	fw_v?: string;
	fhi?: number;
}